finished modifying tests. Though, still needed separate files. Will be explained in the meeting
