--- src/test/java/il/ac/bgu/cs/bp/bpjs/context/Tests/testingSingleEntityQuery.java
@@ -1,11 +1,0 @@
-package il.ac.bgu.cs.bp.bpjs.context.Tests;
-
-public class testingSingleEntityQuery extends SampleTest {
-
-//    public testingSingleEntityQuery() {
-//        super.afterUpdatingToWholeDBQuery = false;
-//    }
-
-}
-
-
