--- src/test/java/il/ac/bgu/cs/bp/bpjs/context/Tests/testingWholeDBQuery.java
@@ -1,11 +1,0 @@
-package il.ac.bgu.cs.bp.bpjs.context.Tests;
-
-/**
- * The (Hard coded) values we compare our results to are based on the results of the 0.6.6 version of BPjs.
- */
-public class testingWholeDBQuery extends SampleTest {
-//    public testingWholeDBQuery() {
-//        super.afterUpdatingToWholeDBQuery = true;
-//    }
-
-}
