+++ src/test/resources/TestingResources/HotCold/bl.js
