+++ src/test/resources/TestingResources/HotCold/dal.js
