+++ src/test/resources/TestingResources/TestCases/activeOnlyIfInContext.js
