+++ src/test/resources/TestingResources/TestCases/checkingPriority.js
