+++ src/test/resources/TestingResources/TestingWholeDbQuery/HotCold.js
