+++ src/test/resources/TestingResources/TestingWholeDbQuery/checkingPriority.js
